edi-automation/requirements.txt
# Selenium for web automation
selenium>=4.18.0
webdriver-manager>=4.0.0

# Certificate handling
cryptography>=42.0.0
pyOpenSSL>=24.0.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Data handling (최신 버전 사용 - Python 3.13 호환)
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0

# Logging
loguru>=0.7.0

# HTTP requests (for API calls if needed)
requests>=2.31.0

# Date/Time handling
python-dateutil>=2.8.0

# For handling Korean encoding
chardet>=5.2.0
